rename modules. clean up gradle files
