+++ app/build.gradle
@@ -18,15 +18,12 @@ configurations.all { config ->
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "com.android.support:appcompat-v7:$android_support_lib"
     implementation 'com.android.support.constraint:constraint-layout:1.1.0'
     implementation "com.android.support:recyclerview-v7:$android_support_lib"
 
-    implementation 'com.jakewharton.threetenabp:threetenabp:1.1.0'
-
     implementation project(':data-layer')
-    implementation project(':filmeddomain')
 
     //Leak Canary
     debugImplementation "com.squareup.leakcanary:leakcanary-android:$leak_canary_version"
