+++ app/src/main/java/me/joshshin/filmed/adapters/MoviesAdapter.kt
@@ -10,13 +10,13 @@ import android.widget.Toast
 import com.squareup.picasso.Picasso
 import me.joshshin.filmed.R
 import me.joshshin.filmed.adapters.MoviesAdapter.MoviesViewHolder
-import me.joshshin.filmeddomain.dataProvider.DataCallback
-import me.joshshin.filmeddomain.dataProvider.DataCallbackException
+import me.joshshin.domain.dataProvider.DataCallback
+import me.joshshin.domain.dataProvider.DataCallbackException
 import kotlinx.android.synthetic.main.movie_item.view.*
 import me.joshshin.filmed.BuildConfig
 import me.joshshin.datalayer.MoviesProvider
 import me.joshshin.datalayer.network.FilmedApiConstants.BASE_IMAGE_URL
-import me.joshshin.filmeddomain.models.FilmedMovie
+import me.joshshin.domain.models.FilmedMovie
 
 /**
  * Created by Josh Shin on 4/15/18
