+++ domain/build.gradle
@@ -1,0 +1,3 @@
+apply plugin: 'com.android.library'
+
+apply from: '../gradle/common.gradle'
