+++ domain/src/main/res/values/strings.xml
