--- filmeddomain/build.gradle
@@ -1,8 +1,0 @@
-apply plugin: 'com.android.library'
-
-apply from: '../gradle/common.gradle'
-
-dependencies {
-    api 'com.jakewharton.threetenabp:threetenabp:1.1.0'
-    implementation "com.squareup.retrofit2:converter-gson:$retrofit_version"
-}
